refactor: high-level tests
